requirements.txt
elasticsearch
pandas
sodapy
redis
kafka-python
flask